+++ src/components/ui/IconBtn.js
@@ -1,15 +1,14 @@
 import React from 'react';
 import IconButton from 'material-ui/IconButton';
 
-
 function IconBtn(props) {
-    // let disabled = (props.disabled) ? 'disabled={props.disabled}' : '';
+    let iconClass = 'fa fa-' + props.iconClass || 'fa ';
     let style = props.style || {};
     style.height = '';
     style.padding = 0;
 
     return (
-        <IconButton onClick={props.onclick} iconClassName="fa fa-github" style={style} />
+        <IconButton onClick={props.onclick} iconClassName={iconClass} style={style} />
     );
 }
